Split sync back into fetch & commit
Single method approach was never going to work. In order to build an
event, the step parent from the remote must be available.

With separate fetch/commit methods, fetch will have to be built to be
idempotent, so that if it is called multiple times without a commit,
it returns the full fetch result. This should be possible using a
TreeWalk from the remote tracking branch with HEAD/self marked as
uninteresting.

This commit also changes the names of various methods/classes to match
the old approach. E.g.
* SyncResult -> FetchResult
* SyncEvent -> FetchedEvent